Приложение переведено на Spring Boot 4 и Java 21. Без тестов.

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -1,7 +1,7 @@
 create sequence hibernate_sequence start 1 increment 1;
 
 create table message (
-    id int8 not null,
+    id bigserial not null,
     filename varchar(255),
     tag varchar(255),
     text varchar(2048) not null,
@@ -15,7 +15,7 @@ create table user_role (
 );
 
 create table users (
-    id int8 not null,
+    id bigserial not null,
     activation_code varchar(255),
     active boolean not null,
     email varchar(255),
